requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
fpdf2>=2.7.0
Pillow>=9.0.0
python-pptx>=0.6.21
